payroll_positions_added

+++ routes/api.php
@@ -4,6 +4,8 @@ use App\Http\Controllers\Api\AuthController;
 use App\Http\Controllers\Api\DepartmentController;
 use App\Http\Controllers\Api\EmployeeController;
 use App\Http\Controllers\Api\ModeratorController;
+use App\Http\Controllers\Api\PayrollController;
+use App\Http\Controllers\Api\PositionController;
 use App\Http\Controllers\Api\ProfileController;
 use App\Http\Controllers\Api\RoleController;
 use Illuminate\Http\Request;
@@ -23,46 +25,49 @@ use Illuminate\Support\Facades\Auth;
 |
 */
 
-// Admin Routes
-Route::middleware(['auth:sanctum', 'checkRole:Admin' ])->group(function () {
-    // Routes that require the 'admin' or 'moderator' role
-    Route::get('/profiles', [ProfileController::class, 'index'])->name('profile-list');
-    Route::get('/profiles/create', [ProfileController::class, 'create'])->name('create-profile');
-    Route::post('/profiles', [ProfileController::class, 'store'])->name('store-profile');
-    Route::get('/profiles/{profile}', [ProfileController::class, 'edit'])->name('edit-profile');
-    Route::put('/profiles/{profile}', [ProfileController::class, 'update'])->name('update-profile');
-    Route::delete('/profiles/{profile}', [ProfileController::class, 'destroy'])->name('delete-profile');
-    Route::get('/departments', [DepartmentController::class, 'index']);
-    Route::post('/departments', [DepartmentController::class, 'store']);
-    Route::get('/departments/{department}', [DepartmentController::class, 'show']);
-    Route::put('/departments/{department}', [DepartmentController::class, 'update']);
-    Route::delete('/departments/{department}', [DepartmentController::class, 'destroy']);
- 
-});
-Route::middleware(['auth:sanctum', 'checkRole:Moderator'])->group(function () {
 
-    Route::get('/departments', [DepartmentController::class, 'index']);
-    Route::post('/departments', [DepartmentController::class, 'store']);
-    Route::get('/departments/{department}', [DepartmentController::class, 'show']);
-    Route::put('/departments/{department}', [DepartmentController::class, 'update']);
 
- 
- 
-    // Routes that require the 'admin' or 'moderator' role
-    
-  
-});
-Route::middleware(['auth:sanctum', 'checkRole:Employee'])->group(function () {
-    // Routes that require the 'admin' or 'moderator' role
-  
 
- 
-});
-Route::apiResource('roles', RoleController::class);
-
-// Role Routes
 
 
 // Authentication Routes
-Route::post('/auth/login', [AuthController::class, 'loginUser']); // Endpoint for user login
+// Endpoint for user login
 Route::post('/auth/register', [AuthController::class, 'createUser']); // Endpoint for user registration
+Route::post('/auth/login', [AuthController::class, 'LoginUser']); // Endpoint for user registration
+
+Route::apiResource('roles', RoleController::class);
+
+// Group for authentication middleware
+Route::middleware(['auth:sanctum'])->group(function () {
+    // Profiles Routes
+    Route::get('/profiles', [ProfileController::class, 'index'])->middleware(['checkRole:Admin,Moderator']);
+    Route::get('/profiles/create', [ProfileController::class, 'create'])->middleware(['checkRole:Admin,Moderator']);
+    Route::post('/profiles', [ProfileController::class, 'store'])->middleware(['checkRole:Admin,Moderator']);
+    Route::get('/profiles/{profile}', [ProfileController::class, 'show'])->middleware(['checkRole:Admin,Moderator']);
+    Route::get('/profiles/{profile}/edit', [ProfileController::class, 'edit'])->middleware(['checkRole:Admin,Moderator']);
+    Route::put('/profiles/{profile}', [ProfileController::class, 'update'])->middleware(['checkRole:Admin,Moderator']);
+    Route::delete('/profiles/{profile}', [ProfileController::class, 'destroy'])->middleware(['checkRole:Admin']);
+
+    // Departments Routes
+    Route::get('/departments', [DepartmentController::class, 'index'])->middleware(['checkRole:Admin,Moderator']);
+    Route::get('/departments/create', [DepartmentController::class, 'create'])->middleware(['checkRole:Admin,Moderator']);
+    Route::post('/departments', [DepartmentController::class, 'store'])->middleware(['checkRole:Admin,Moderator']);
+    Route::get('/departments/{department}', [DepartmentController::class, 'show'])->middleware(['checkRole:Admin,Moderator']);
+    Route::put('/departments/{department}', [DepartmentController::class, 'update'])->middleware(['checkRole:Admin,Moderator']);
+    Route::delete('/departments/{department}', [DepartmentController::class, 'destroy'])->middleware(['checkRole:Admin']);
+
+    // Positions Routes
+    Route::get('/positions', [PositionController::class, 'index'])->middleware(['checkRole:Admin,Moderator']);
+    Route::get('/positions/create', [PositionController::class, 'create'])->middleware(['checkRole:Admin,Moderator']);
+    Route::post('/positions', [PositionController::class, 'store'])->middleware(['checkRole:Admin,Moderator']);
+    Route::get('/positions/{position}', [PositionController::class, 'show'])->middleware(['checkRole:Admin,Moderator']);
+    Route::put('/positions/{position}', [PositionController::class, 'update'])->middleware(['checkRole:Admin,Moderator']);
+    Route::delete('/positions/{position}', [PositionController::class, 'destroy'])->middleware(['checkRole:Admin']);
+
+    Route::get('/payrolls', [PayrollController::class, 'index'])->middleware(['checkRole:Admin,Moderator']);
+    Route::get('/payrolls/create', [PayrollController::class, 'create'])->middleware(['checkRole:Admin,Moderator']);
+    Route::post('/payrolls', [PayrollController::class, 'store'])->middleware(['checkRole:Admin,Moderator']);
+    Route::get('/payrolls/{payroll}', [PayrollController::class, 'show'])->middleware(['checkRole:Admin,Moderator']);
+    Route::put('/payrolls/{payroll}', [PayrollController::class, 'update'])->middleware(['checkRole:Admin,Moderator']);
+    Route::delete('/payrolls/{payroll}', [PayrollController::class, 'destroy'])->middleware(['checkRole:Admin']);
+});
